Add 'capacity' function
