+++ tests/filtration_tests/test_grabber.py
@@ -5,10 +5,10 @@ import pytest
 from lz.filtration import grabber
 from lz.hints import (Domain,
                       Predicate)
+from lz.iterating import capacity
 from lz.replication import duplicate
 from tests.utils import (are_iterables_similar,
-                         are_objects_similar,
-                         capacity)
+                         are_objects_similar)
 
 
 def test_basic(iterable: Iterable[Domain]) -> None:
